requirements.txt
fastapi==0.95.1
openai==0.27.6
pandas==2.0.1
pydantic==1.10.12
PyPDF2==3.0.1
python-dotenv==1.0.0
qdrant_client==1.4.0
sentence_transformers==2.2.2
